Split fixtures into conftest
